Break up headers a bit more, added XML doc comments to API declarations
