+++ code/include/playfab_c/PlayFabClientDataModels_c.h
@@ -499,14 +499,6 @@ struct PlayFabPlayerProfileResult
     PlayFabPlayerProfileModel* playerProfile;
 };
 
-struct PlayFabLoginResult
-{
-    // TODO figure out which fields to expose here
-    PlayFabUserHandle userHandle;
-    time_t* lastLoginTime;
-    bool newlyCreated;
-};
-
 struct PlayFabCreateSharedGroupRequest
 {
     const char* sharedGroupId;
